Attempt to prevent System.exit
